Change errors string output to null on eval
Glass only checks that an error string is null on the output of eval, not
that it is null or empty. I fixed this in glass but it got lost when I
moved the changes from git to tfs in the OOB branch. This change makes
MIEngine work with glasses eval.