Support 32-bit apps on 64-bit Android devices
Address the following problems we had when targetting a 64-bit Android
device:

- Our ABI detection logic fell down because the device's ABI is arm64
instead of one of the arm values. Changed the code to check the
'abilist' property as well.

- At least on Galaxy S6 thet we tested, run-as was adding extra debug
spew which broke our processing of the output from pwd. Modified this
code to ignore the extra spew.

- The 32-bit app_process is now called 'app_process32' and at least on
this device, ADB wasn't handling the redirect