Improve error message for invalid android/iOS launch options
This checkin fixes a bug where if there was something wrong with the
launch options, the user would see a message with a duplicated 'Launch
options string provided by the project system is invalid'. Ex:

---------------------------
Microsoft Visual Studio
---------------------------
Unable to start debugging. Launch options string provided by the
project system is invalid. Launch options string provided by the
project system is invalid. Required attribute 'LaunchActivity' is
missing.
---------------------------
OK
---------------------------

To fix this, I added a new InvalidLaunchOptionsException exception
class. This way in LaunchOptions.cs:668, the code can tell if the
exception is already an InvalidLaunchOptionsException, in which case it
doesn't need to wrap it a second time.