Fix official build to drop files
The official build is no longer dropping the build files to the build share. The problem is that this used to work for setting 'OutDir' but BuildAndTest.proj forces OutDir to a particular value. The other problem we had was that the output directory was getting filled with VS reference assemblies.

To resolve this, I created a new 'DropFiles' target in MIDebugPackage.csproj which takes care of copying the files.

I also marked our various VS reference assemblies with '<Private>False</Private>' which mostly fixed our build to stop dropping VS assemblies into the output directory.