Fix races in Breakpoint test
The breakpoints test was not declaring IDebugBreakpointBoundEvent2 as expected, which meant that is was non-deterministic which command the event would arrive in. This changes the test to wait for the bound event before continuing. I also remove some of the duration elements to minimize the differences between the test script and session log.