Add device udid to appRemotePath call for vcremote
We need to send the device udid to both the seutp for debugging call and
the appremotepath call. In a perfect world, we would make this a global
setting in the header of all vcremote packets, but that fix will have to
wait for another day.

Additionally, fix a URI issue in setupForDebugging where the second '?'
needed to be an &.