I fought the build, and I won (package management fix)
Summary: With this checkin, I am changing how we restore packages in
the MI Engine, and also, on a much more minor note, how we compile our
contract IL.

Problem: I couldn't figure out any normal way to set the package output
directory when NuGet restores of our project.json projects in the IDE.
I didn't like this, but could live with it if this was the only
problem. But on the PR machines packages were getting restored to some
unknown directory, so it wasn't working at all.

Solution:
1. Move package restore to the same spot in both the IDE build and the
build.cmd build by creating 'prebuild.csproj', and making all projects
depend on this project.

_NOTE_ when adding a new project, we need to add this new dependency.
To do so, bring up the project context menu and invoke:
  Build Dependencies->Project Dependencies

This will add a new item in the .sln file to declare the dependency, so
as long as we review .sln changes, we should be good.

2. Have prebuild.csproj set the package output switch (which actually
is respected)

3. Set the NuGetPackagesDirectory property in all projects, so projects
know where to look for their dependencies.

4. For good measure, I moved the .il compilation from MICore to
prebuild.csproj since that is really where it belongs.