Create Microsoft.MICore.XmlSerializers.dll
After an investigation, I determined that there is no practicle way to use sgen code with CoreCLR - the types aren't in a contract, so therefore they are subject to change. So this makes changes to run sgen in a more traditional way where the sgen code is in its own dll. This way MICore can stay portable but we will still get the benefits of a static XML deserializer in the VS IDE, which is where we care about it anyway.