Fix InvalidOperationException in JDbg 'OnDisconnect'
This fixes an issue we saw come in through Watson. On disconnect, JDbg
attempts to abort all existing WaitingOperations. The abort inside of a
foreach loop was causing the dictionary to be altered, which causes an
InvalidOperationException. Fix this by copying each
WaitingOperationDescriptor into a temporary list, clearing the main list,
and then aborting each one out of the temporary list.