Remove global DebuggedProcess variable
This does not completely pave the way for multiple process debugging from
MIEngine, but it sets the stage. This removes the evil global variable
storing the one and only DebuggedProcess object.

TODO: remove the DebuggedProcess property from AD7Engine and allow for
multiple DebuggedProcess objects to be created and managed.