Fix UseUnixSymbolPaths for non-local launch options
Back before update 1, UseUnixSymbolPaths would return true for any non-local target. This was wrong for LocalLaunchOptions once we added support for running on non-Windows, so we fixed that. But we broke the scenario for non-LocalLaunchOptions. So this restores that code path. This still isn't completely correct since technically you could use Pipe/Tcp launch options to connect to Windows. But we don't have any scenarios yet where we want to do that.