Add packaging target
With this change we will package the MIEngine bits for use with clrdbg as
part of the lab builds. An initial nuspec was added that includeds only
the assemblies needed for VSCode. Also, version.settings.targets was
abstracted out of miengine.settings.targets so it could be included in
buildandtest.proj in addition to all the projects under src.