Move PackageNuGet target to its own file, add package to drop folder
With the last commit, the nuget package was being created but not ending
up in the drop folder so it was not getting placed on the drop server.
This commit moves the target into it's own file and includes it in
MIDebugPackage.csproj so that we can make use of the DropFiles properites.