Add support for non-absolute paths in 'ExePath'
The .NET CLI is making changes where we want to tell clrdbg to launch 'dotnet'. This makes some MIEngine tweaks so that we let clrdbg/gdb/lldb decide if the passed in program path is valid instead of having the MIEngine pre-validate it, and then do a reasonable job handling failures from the underlying debugger.