Add launch option for controlling external console
This checkin adds a new launch option for controlling if an external console should be
used and modifies the external console support so that it can be used with CLRDBG.

Remaining work:
* I am currently sometimes seeing an access violation which leads to a hang inside of
clrdbg during shutdown. This is inside of the JIT, so I want to try upgrading my CoreCLR
before investigating more.
* Support OSX