LocalTransport: Stop corrupting the PATH environment variable
LocalTransport had some code meant for working around a bug with GDB on
Windows where the GDB directory needed to be on the path. However,
it was doing this in all scenarios.

It's possible that there are other scenarios where this code might be useful,
but in that case it needs to use the ':' separator on UNIX instead of
corrupting the path. For now, I just went with the easy fix of turning it off
except for where we know we need it.