Fix InstallToVSCode for breaking CLI change
The CLI team previously indicated that a native host executable will not be dropped
unless the project.json explicitly indicates that the application wants to be
platform depentant. But builds from a few days ago were actually behaving nicer than
the way I told it should work.

Now things are actually behaving this way. So I am fixing up InstallToVSCode to generate
project.json on the fly so we can write out a runtime id.