Use '-gdb-set new-console' on Windows
On Windows, GDB has an option to launch using a new console. CLRDBG was changed to use the same syntax. This changes the MIEngine to use this syntax, and I removed the 'if clrdbg' check.

Note that I don't believe that LLDB supports this option yet. But this shouldn't blow up as we are ignoring failures.