Add support for cygwin on windows
Cygwin maps files from windows paths to unix style
paths. It has a utility called cygpath that will
automatically do the conversion. Calling that utility
proved to be quite difficult because cygwin will fail
to initialize stacks if a 64bit app is connected to a
32 bit console. Under vscode, the 32 bit console is
inherited from vscode. This works around this by launching
that utility with process_detached to force a new console.
Note that doing so requires pinvoking to createprocess
directly since the Process object doesn't expose the
raw flags.

cr feedback