Monitor the shell pid and remove the exit fifo
The exit fifo was causing issues in the case that MIEngine had already exited
but the shell started executing the EXIT trap. In that case, it would block on
echoing 'exit' to a fifo since there were no readers.

Use getpgid instead of kill -0 to find if a process is running