Implement IDebugEngine3.SetEngineGuid
This checkin implements IDebugEngine3 in the MIEngine. The reason for this is
to provide support for SetEngineGuid, which allows registering the MIEngine
for an alternate engine GUID without needing to change the code.

The value of this is that, for the SSH port supplier, we can use a different
engine GUID to differentiate between LLDB/GDB/clrdbg and thus offer multiple
types of code debugging using the same gestures that the attach to process
dialog already knows (engine selection).