Previous console work for GDB does not work with LLDB
Jackson made changes in LLDB to support -gdb-set new-console on in LLDB to
mimic the way we do console/terminal windows in Windows.

NOTE: the changes from the lldb side are also required

This also includes a change to ignore SIGHUP at the debugger level and
pass it to the debuggee. This happens when the console/terminal window is
closed.