SSHLaunchOptions support
This checkin adds support for using XML launch options to initiate a launch. This will allow us to support a not-quite-so-off-road launch experience for Dev15 until we get real project system support sometime later.

Note that in this scenario we don't actually use the SSH port supplier as a real port supplier. It is just a transport and as far as AD7 is concerned, we are still using default transport.