The change that obtained the target architecture in the miengine introduced a subtle race condition as executing this command requires break mode, but the module loads may not have finished. This caused a complete failure of the attach / launch sequence in multi-threaded apps on ubuntu 16.04. However, given the nature of this race, I believe it likely is possible to repro on other platforms.
My fix for this is to delay that operation until the first break
(which occurs right after statically linked shared libraries)
by injecting it in the front of the break action queue. note
that the existing function for adding to that queue would not
work because the connected status wasn't set yet.