Improve handling of debugger aborts
This checkin improves the error message that we show when GDB exits or we otherwise loose the connection to the target process.

This also fixes a hang that could happen if we lost our connection after DebuggedProcess.Initialize, but before DebuggedProcess.ResumeFromLaunch. The issue was that nothing would send the program destroy event.