Stop calling CheckModules for lldb/clrdbg (#542)
MIEngine had a method that we use for GDB to find the list of loaded modules. This method wasn't intended to be called for lldb/clrdbg, but there was one race where it could happen - if a stopping event was received before the SDD continued program create we will call CheckModule.

This adds a guard to fix this.

This should resolve https://github.com/OmniSharp/omnisharp-vscode/issues/1035.