Fix NullRef in SSH port supplier (#575)
There's a race condition between a SSH command completing normally and
being disposed - the member for the command is set to null when the
command is disposed, but handler for the command's "Finished" event
still references the member.  The fix is to have the "Finished" event
handler access the command via the "sender" instead.