Update package references
The MIEngine was still compiling against incredibly ancient packages. This switches it over to something recent - the packages that came out at the same time as .NET Core 1.0 RTW.

The primary IL level difference of this change is to increment the reference assembly version of the following in the non-'Desktop' configurations.

    System.Xml.XmlSerializer
    System.IO.FileSystem
    System.Net.Http
    System.IO.FileSystem.Primitives

In addition, this changes:
* The Desktop configuration to stop dropping a bunch of facade assemblies that it doesn't seem like it actually needed.
* Stops having the IOS launcher need to partially depend on facade assemblies. To do this I added an MIEngine definition of RemoteCertificateValidationCallback.