Fix termination on MinGW/Cygwin by calling Windows API to terminate.
MinGW and Cygwin do not allow Async break to occur without user
intervention. The problem is with -exec-abort (stop debugging) while in
run mode requires us to do an async break so that we can abort. Since we
can't do that without additional user intervention, we will terminate
the child process instead.