Change -stack-list-arguments call to be more performant in some cases
Made change to call -stack-list-arguments with --no-values when the
value flag is not specified. This improves perf for large objects in stack walks but it does add an
extra call per variable to -var-create and -var-delete to get the type.
The calls to -var-create/-var-delete are faster overall than the
-stack-list-arguments with --simple-values.

If both value and types are requested, we fall back to requesting
-stack-list-arguments with simple values which will return types and
values.