Use -- instead of -x for gnome-terminal
The -x option is deprecated, so us -- instead.

In Ubuntu 18.04, using the `-x` option results in the program being launched in a new tab in an existing terminal. This terminal is not brought to the foreground, nor is the tab activated, which makes a bunch of work for the user to find where there program is. Using the `--` option, on the other hand, results in the program opening in a new terminal and the new terminal is brought to the foreground as one might expect.