Fix MIEngine file version
This changes MIEngine so that the file version of all of the files built in the repo is 16.5.* instead of 14.0.*. This is important because our date-based version numbers overflowed 2^16 due to our lack of dilegance in updating it, and the project wouldn't build.

I left the assembly version as 14.0 for compatibility.
This also fixes the build of our various unit test projects. I don't know how this wasn't a problem before, but now it works.

Testing: Verified SSH attach to process