Fixes for environment entries used for RunInTerminal Request (#981)
It was accidently using the debuggee's environment variables to also be
used for the RunInterminal Request. Also as @Smit-tay pointed out, we
should probably be using /usr/bin/env in the script and a full path to
sh in the script execution when calling RunInTerminal.

Added some fixes and a comment. We may want to support RunInTerminal
Environment Variables in the future so I did not remove the plumbing for
it, but for now I'm sending an empty object.