Remove project.json files and clean up build (#982)
This commit removes the "project.json" files that were still being used
in some build configurations.  Where necessary, NuGet package references
are moved to the relevant "packages.config" files.

This commit also gets rid of the confusingly-named `IsCoreClr` MSBuild
variable and the `CORECLR` compile-time flag.  These are replaced with a
new `IsXPlat` MSBuild variable and `XPLAT` compile-time flag.  Note that
the meaning of this flag is inverted - code that shouldn't be included in
the cross-plat build used for the VS Code C++ extension used to be
excluded via `#if CORECLR`, while the new exclusion would be `#if
!XPLAT`.

`IsXPlat` and `XPLAT` are both defined when building one of the
`Desktop.*` build configurations and are undefined for `Debug`,
`Release`, and `Lab.*`.