Support ClipboardContext for Evaluate Request (#1009)
* Support ClipboardContext for Evaluate Request

Users who want to copy the full string from the UI will use "copy value"
from the EE windows. This will send an EvaluateRequest with the context
of Clipboard.

This PR noticies the context in OpenDebugAD7 and uses IDebugExpressionDAP
with DAPEvalFlag ClipboardContext set for MIEngine to query.
If it is set, MIEngine will save the previous value of 'show print
elements', set it to unlimited, and reset it back to the original value
after it evaluates the variable. Then it will provide the full string
back to the user to copy.