[Docker] Fix remote containers list, add OutputWindow error support and Telemetry (#1017)
In the Container Picker Dialog, it uses Name and ShortId. However, if
Id is null, ShortId tries to get the length of the Id and this causes a
NullRefException.

We also are adding null containers for GetRemoteDockerContainers. Added
a null check. 

Adding a null check on Id so it just returns null.

Added support for logging errors to the output window.

Added telemetry to count for errors in Docker process listings