Added Platform field to the container header and expanded view in the "Select Docker Container" window (#1021)
Added Container OS/Platform information in the Docker container picker view. 

It will query `docker version` to get the Server's OS, however is LCOW is enabled and the user set the docker daemon to use Windows Containers. 

We will query using `docker inspect` to see if it acutally is running Windows or Linux on the containers.