Fix SourceFileMap in MICore
Added PathNormalization methods in PlatformUtilities that is both used
in MICore and MIDebugEngine.

This change will ensure all paths created for sourceFileMap are
normalized to the client. E.g. Windows '\' if on Windows and '/' if on
macOS or Unix.

When submitting for breakpoints in MIEngine, it will normalize compiler
sources to use Unix styled paths for GDB.