Fix XUnit Test Analyzer Errors
xUnit2000: The literal or constant value <..> should be passed as
the 'expected' argument in the call to 'Assert.Equal(expected, actual)'

xUnit2007: Do not use typeof(<Type>) expression to check the type.

xUnit2009: Do not use Assert.True() to check for substrings.

xUnit1013: Public method 'VerifyCoreApisPresent' on test class
'BasicLaunchOptionsTests' should be marked as a Fact.