Support Empty/Null strings in 'args' for json
This PR handle empty/null strings in the args field for the launch.json.
QuoteArgs will quote empty strings if requested and we will skip over
null strings since it is not possible to pass NULL into a program's
argument.

However, launch.json's 'args' array does support it so we are
handling it so it does not crash OpenDebugAD7.