Support non-IDebugMemoryBytesDAP engines
There can be cases where the engine loaded is not the latest AD7Engine.
In GetMemoryContext, we do a hard cast to IDebugMemoryBytesDAP. This
will error with "Specified cast is not valid" if the engine does not
implement IDebugMemoryBytesDAP and the engine is not the latest
Microsoft.MIDebugEngine.

This PR converts the hard cast to be a `is` check. If the engine is not
IDebugMemoryBytesDAP, it will return E_NOTIMPL.