Support "format" field on "stackTrace" request
Adds support for the "value formatting" feature of DAP, which allows flags
to be passed to control formatting of stack frames, as well as variables
and evaluation results.

Note that while this commit wires up the DAP formatting flags to the
relevant AD7 calls, MIEngine currently doesn't appear to respect the
"radix" value passed, so the flag to control hex vs decimal formatting
doesn't work.