Fixing Cygwin PathMapping with SourceFileMap
When migrating sourceFileMap to MIDebugEngine in the XML Conversion PR,
we are now doing the call to 'cygpath' after we attempt to do the source
file mapping.

This PR moves this call before we do the source file map for frame's
files.

This also removes the DebuggedProcess in AD7DocumentContext and cleans
up the constructors and those who initalize the AD7DocumentContext.