Use SIGTRAP to interrupt a process instead of SIGINT (#1090)
Because of https://bugzilla.kernel.org/show_bug.cgi?id=9039, when
a program blocks a signal because it uses sigwait() or signalfd(),
GDB isn't notified of the signal via ptrace and cannot intercept it.

For MIEngine, this means that whenever MIEngine attaches to a process
using signalfd() or sigwait() to process SIGINT, any SIGINT signal
sent by MIEngine to interrupt gdb is delivered to the process instead
which usually handles the signal by exiting.

As a workaround, let's use SIGTRAP instead. SIGTRAP is explicitly
defined as a signal to inform the debugger that some condition has
occurred that it has requested to be informed of. As such, it is a
lot less likely to be blocked by programs compared to SIGINT. Because
it isn't blocked most of the time, gdb will be able to intercept it
and it won't be delivered to the program.

```
SIGTRAP
   ; The SIGTRAP signal is sent to a process when an exception (or trap) occurs: a condition that a debugger has requested to be informed of — for example, when a particular function is executed, or when a particular variable changes value.
```

Fixes https://github.com/microsoft/vscode-cpptools/issues/6704.

Co-authored-by: Andrew Wang <waan@microsoft.com>