Support GetProcessArchitecture and read ps flags
This PR adds GetProcessArchitecture in IDebugUnixProcess for SSHDebugPS
and obtains the flags in ps.

For most processes, it will be the same as the system architecture from
'uname -m'.
However, on Apple M1 Machines. we need to determine if the process is being
emulated or not.
We determine this by checking if P_TRANSLATED (0x20000) on OSX ARM64
is enabled on the process, it is under x86_64 emulation.