Fix LogPoint Issue
Customer reported in v1.6.0 that LogPoints stopped working.
This was due to IDebugBreakpointEvent2.EnumBreakpoints returning the
same IEnumDebugBoundBreakpoints2.

The fix was to save IDebugBoundBreakpoint2[] in the event and return a
new IEnumDebugBoundBreakpoints2 object per call.

Added a test for logpoints and fixed up ConsoleEvent to be OutputEvent
since thats what it should actually be.