Fix GitHub Action Test runs
This PR revert "Remove downgrade to GNU toolchain (#1298)" since
machines seem to be flaky on which versions of GDB we get.

This reverts commit 8d70a37d2243862eccba70a6226bf684708871c7.

We also introduce retry logic for dotnet test on flaky tests.

Also does a null check on writer.close since that seems to be the cause
of most tests failing when disconnecting.