Fix deleting breakpoints created as pending
This PR fixes https://github.com/microsoft/vscode-cpptools/issues/9095.
The issue is that breakpoints that are created on debug session start
have not had their module loaded yet and BoundBreakpoints will be null
and with the current check we will return E_FAIL and never cache the
breakpoint.

The original usage of this was to capture errors for data breakpoints,
but they will come through as BreakpointEvents instead of the response
of SetBreakpoints / SetDataBreakpoints.