Address visualizing past 1000 Array Items
GDB seems to have a limit of 1000 child items that can be returned when
using (T[<Size>])*(<ValuePointer>).

This PR modified the request to be (T[<At most 50>])*(<ValuePointer> +
StartOffset).

Also added a test case to see if we can see past 1000 by testing with an
ArrayItem visualized type with 2000 elements.