[#1370] Change the type of breakpoint ids from `ulong` to `long`
GDB represents a breakpoint number as an `int`. See the definition of
`struct breakpoint` in breakpoint.h, which includes:

    /* Number assigned to distinguish breakpoints.  */
    int number = 0;

GDB uses negative breakpoint numbers to represent "internal"
breakpoints. See the "Setting Breakpoints" documentation [1] which
says:

    GDB itself sometimes sets breakpoints in your program for special
    purposes, such as proper handling of longjmp (in C programs).
    These internal breakpoints are assigned negative numbers, starting
    with -1; "info breakpoints" does not display them. You can see
    these breakpoints with the GDB maintenance command "maint info
    breakpoints"

Internal breakpoints can also be created in Python [2], for example:

    python gdb.Breakpoint("function_name", internal=True)

When an internal breakpoint is hit, MIEngine must not crash, so this
commit changes the representation of breakpoint numbers from unsigned
to signed integers. This avoids a `System.OverflowException` from the
call to `Convert.ToUint32` in `TryGetExceptionBreakpoint`.

[1] https://sourceware.org/gdb/onlinedocs/gdb/Set-Breaks.html
[2] https://sourceware.org/gdb/onlinedocs/gdb/Breakpoints-In-Python.html