Fix NPE with watching NatvisSection
This PR fixes the variable used for the RegistryMonitor on a fresh VS
instance.

If the private registry does not have a NatvisDiagnostic section, it
will watch for its creation then delete this watcher.
However when creating the new watcher it would watch on a null section
instead of the new natvis section.

This PR fixes that issue.