MIEngine: Fix multi-breakpoint gdb/mi information parsing
This PR fixes gdb/mi breakpoint information parsing for breakpoints
that bind to multiple locations, which have a '<MULTIPLE>' in the addr field.
According to https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Breakpoint-Information.html#GDB_002fMI-Breakpoint-Information
the locations field contains the list of locations for this breakpoint type.
This list of locations is used then to create the bound breakpoints for the specific breakpoint.

Signed-off-by: Andria Pazarloglou <androniki.pazarloglou@intel.com>