Only use ps 'flags' for macOS
This PR adds in an operating system check via uname to determine if we
should run the PS command with the 'flags' flag. This feature was
originally used to determine the architecture of a process for macOS M1
machines that can run as x64 or as arm64.

AD7Process only uses the flags field in macOS so this PR updates the
command sent to only use it on macOS machines.

Other changes:
- Updated the setup.csx script to update VS with VS.list instead of
  VS.Codespaces.list
- Updated tests to use the new PSOutputParser
- Added a macOS test