Fix SSH attach on a clean machine (#1444)
Attaching to a C# process running on a brand-new Linux VM was failing with an error about being unable to create or access the `.vs-debugger` folder on the remote host.  After investigating, this is due to a behavior change in liblinux - previously, the call to `IRemoteFileSystem.Stat()` would fail if the file didn't exist, but it now returns a valid object that you can query to determine that the file doesn't exist.  I updated the logic accordingly.

Workaround: manually create the `.vs-debugger` folder on the remote host.  Once it exists, the remote debugger will deploy successfully.