Fix name for enum keys in a map holding structs
When a map has enum as keys and any type of struct as values, the enum value that was used for the key would be replaced as if it was base class.

As the VariableInformation class has a constructor that receives the names directly from the map, use that information to not allow the name override.